// src/app/[locale]/(private)/module/k7-form/components/types.ts
export interface UniversityFilterFormValues {
  year?: number;
  facultyId?: number;
  departmentId?: number;
  profileIndex?: number;
}

export interface K7UniversityFilterSelection {
  year?: number;
  facultyId?: number;
  departmentId?: number;
  targetAccountId?: number;
  employeeId?: number;
  position?: string;
}